+++ addon/components/fd-editform-constructor.js
@@ -25,6 +25,8 @@ import FdDataTypes from '../utils/fd-datatypes';
 export default Component.extend({
   layout,
 
+  classNames: ['row'],
+  
   /**
     Service for managing the state of the application.
 
@@ -726,7 +728,7 @@ export default Component.extend({
     let types = [];
 
     let clazz = view.get('class');
-    let definition = view.get('definition');
+    let definition = view.get('definitionArray');
     let inheritances = this.get('model.inheritances');
     let associations = this.get('model.associations');
     let aggregations = this.get('model.aggregations');