+++ CHANGELOG.md
@@ -3,18 +3,39 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [0.13.0-beta01]
+## [0.14.0-beta02]
+### Fixed
+* Fix locales
+
+## [0.14.0-beta01]
+### Fixed
+* Fix setButtonStyles
+
+## [0.13.0]
 ### Added
+* Add resize and colors to statechart diagram.
+* Add resize and colors to deployment diagram.
+* Add resize and color to activity diagram.
+* Add table for user access settings.
+* Add storage table.
+* Add rename stage objects function.
 * Add template class resize.
 * Add check errors on all projects page.
 * Add highlight for links.
 * Add links edit forms.
+* Added swimline primitive.
+* Add locks for diagrams and repository objects.
 
 ### Fixed
+* Highlight optimization.
+* Fix rollback new class on diagrams.
+* Fix link role label.
+* Feature update view when deleting class.
 * Fix rollback diagrams.
 * Fix colors in collapsed class.
 * Fix locales in groups.
 * Fix group by system.
+* Fix actors size with proportions.
 
 ## [0.12.0]
 ### Added